requirements.txt
flask
Flask-SQLAlchemy
flask-restful
flask-marshmallow
flask-jwt-extended
Werkzeug
psycopg2-binary